improve embeddings to fix issues with identifying names
I was asking kb about some meetings and noticed it would not return
results for non-Western names. I restructured the way the embeddings
work to split the documents by headers and lowered the chunk size.

The real fix was to structure my meetings notes to make it more clear
what people I was meeting with.